+++ system/libraries/Controller.php
@@ -2962,7 +2962,7 @@ abstract class Controller extends System
 		$this->Database->prepare("DELETE FROM tl_version WHERE tstamp<?")
 					   ->execute((time() - $GLOBALS['TL_CONFIG']['versionPeriod']));
 
-		// Get new record
+		// Get the new record
 		$objRecord = $this->Database->prepare("SELECT * FROM " . $strTable . " WHERE id=?")
 									->limit(1)
 									->executeUncached($intId);
@@ -3077,6 +3077,56 @@ abstract class Controller extends System
 		}
 
 		return $arrReturn;
+	}
+
+
+	/**
+	 * Get the parent records of an entry and return them as string
+	 * which can be used in a log message
+	 * @param string
+	 * @param integer
+	 * @return string
+	 */
+	protected function getParentRecords($strTable, $intId)
+	{
+		// No parent table
+		if (!isset($GLOBALS['TL_DCA'][$strTable]['config']['ptable']))
+		{
+			return '';
+		}
+
+		$arrParent = array();
+
+		do
+		{
+			// Get the pid
+			$objParent = $this->Database->prepare("SELECT pid FROM " . $strTable . " WHERE id=?")
+										->limit(1)
+										->execute($intId);
+
+			if ($objParent->numRows < 1)
+			{
+				break;
+			}
+
+			// Store the parent table information
+			$strTable = $GLOBALS['TL_DCA'][$strTable]['config']['ptable'];
+			$intId = $objParent->pid;
+
+			// Add the log entry
+			$arrParent[] = $strTable .'.id=' . $intId;
+
+			// Load the data container of the parent table
+			$this->loadDataContainer($strTable);
+		}
+		while ($intId && isset($GLOBALS['TL_DCA'][$strTable]['config']['ptable']));
+
+		if (empty($arrParent))
+		{
+			return '';
+		}
+
+		return ' (parent records: ' . implode(', ', $arrParent) . ')';
 	}