+++ system/modules/news/ModuleNewsArchive.php
@@ -173,7 +173,7 @@ class ModuleNewsArchive extends ModuleNews
 			$page = $this->Input->get('page') ? $this->Input->get('page') : 1;
 
 			// Do not index or cache the page if the page number is outside the range
-			if ($page < 1 || $page > ceil($total/$this->perPage))
+			if ($page < 1 || $page > max(ceil($total/$this->perPage), 1))
 			{
 				global $objPage;
 				$objPage->noSearch = 1;