+++ system/modules/frontend/ContentHyperlink.php
@@ -129,7 +129,7 @@ class ContentHyperlink extends ContentElement
 		if ($this->target)
 		{
 			global $objPage;
-			$this->Template->target = ($objPage->outputFormat == 'xhtml') ? ' onclick="window.open(this.href);return false"' : ' target="_blank"';
+			$this->Template->target = ($objPage->outputFormat == 'xhtml') ? ' onclick="return !window.open(this.href)"' : ' target="_blank"';
 		}
 	}
 }