+++ system/modules/frontend/Module.php
@@ -287,7 +287,7 @@ abstract class Module extends Frontend
 					// Override the link target
 					if ($objSubpages->type == 'redirect' && $objSubpages->target)
 					{
-						$row['target'] = ($objPage->outputFormat == 'xhtml') ? ' onclick="window.open(this.href);return false"' : ' target="_blank"';
+						$row['target'] = ($objPage->outputFormat == 'xhtml') ? ' onclick="return !window.open(this.href)"' : ' target="_blank"';
 					}
 
 					$items[] = $row;
@@ -320,7 +320,7 @@ abstract class Module extends Frontend
 					// Override the link target
 					if ($objSubpages->type == 'redirect' && $objSubpages->target)
 					{
-						$row['target'] = ($objPage->outputFormat == 'xhtml') ? ' onclick="window.open(this.href);return false"' : ' target="_blank"';
+						$row['target'] = ($objPage->outputFormat == 'xhtml') ? ' onclick="return !window.open(this.href)"' : ' target="_blank"';
 					}
 
 					$items[] = $row;