+++ system/modules/rep_client/languages/de/tl_repository.php
@@ -93,6 +93,7 @@ $GLOBALS['TL_LANG']['tl_repository']['updatehint']         = 'Status auswählen,
 $GLOBALS['TL_LANG']['tl_repository']['updatingext']        = 'Aktualisierung %s %s Build %s';
 $GLOBALS['TL_LANG']['tl_repository']['validate']           = 'Reparieren';
 $GLOBALS['TL_LANG']['tl_repository']['validatingext']      = 'Prüfung/Reparatur %s %s Build %s';
+$GLOBALS['TL_LANG']['tl_repository']['missingSoapModule']  = 'SOAP-Erweiterung nicht geladen (konfigurieren Sie PHP mit --enable-soap).';
 
 
 /**