+++ system/modules/backend/BackendModule.php
@@ -101,8 +101,6 @@ abstract class BackendModule extends Backend
 		{
 			return null;
 		}
-
-		return $e;
 	}