+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Version 2.11.3 (XXXX-XX-XX)
 ---------------------------
 
 ### Fixed
+Define `BE|FE_USER_LOGGED_IN` in the cron script (see #4099).
+
+### Fixed
 Correctly align the versions drop-down menu (see #4083).