+++ CHANGELOG.md
@@ -5,6 +5,10 @@ Version 2.11.3 (XXXX-XX-XX)
 ---------------------------
 
 ### Fixed
+Convert the "rel" attribute inserted by TinyMCE to a "data-lightbox" attribute
+if it is an HTML5 page (see #4073).
+
+### Fixed
 Uploaded files should now be resized correctly (see #3806).
 
 ### Fixed