+++ CHANGELOG.md
@@ -5,6 +5,9 @@ Version 2.11.3 (XXXX-XX-XX)
 ---------------------------
 
 ### Fixed
+Open popup windows so they are not blocked (see #4243).
+
+### Fixed
 Replaced `is_a()` with `instanceof` in the simplepie plugin (see #4212).
 
 ### Fixed